# src/main/resources/api.yml
openapi: 3.0.1
info:
  title: Open Hospital API
  version: v0
servers:
  - url: http://localhost:8081
    description: Generated server url
paths:
  /api/patientRecord/{id}:
    get:
      tags:
        - patient-record-api
      operationId: getPatientRecord
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PatientRecordDTO'
    put:
      tags:
        - patient-record-api
      operationId: updatePatientRecord
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientRecordDTO'
        required: true
      responses:
        "200":
          description: OK
    delete:
      tags:
        - patient-record-api
      operationId: deletePatientRecord
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
...
  /api/patientRecord:
    post:
      tags:
        - patient-record-api
      operationId: createPatientRecord
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientRecordDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PatientRecordDTO'
...
components:
  schemas:
    DoctorDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        phoneNumber:
          type: string
        address:
          type: string
        email:
          type: string
        avatar:
          type: string
        profession:
          type: string
        latestPatients:
          type: array
          items:
            $ref: '#/components/schemas/PatientDTO'
    PatientDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        opd:
          type: integer
          format: int64
        idp:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        phoneNumber:
          type: integer
          format: int64
        address:
          type: string
        email:
          type: string
        avatar:
          type: string
        bloodGroup:
          type: string
          enum:
            - ZERO_PLUS
            - ZERO_MINUS
            - A_PLUS
            - A_MINUS
            - B_PLUS
            - B_MINUS
            - AB_PLUS
            - AB_MINUS
        notes:
          type: string
        chronicPatient:
          type: boolean
        lastAdmission:
          type: string
          format: date-time
        lastDoctorVisitedId:
          type: integer
          format: int64
        patientRecords:
          type: array
          items:
            $ref: '#/components/schemas/PatientRecordDTO'
        doctorIds:
          type: array
          items:
            type: integer
            format: int64
    PatientRecordDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        patientId:
          type: integer
          format: int64
        doctor:
          $ref: '#/components/schemas/DoctorDTO'
        date:
          type: string
          format: date-time
        typeVisit:
          type: string
        reasonVisit:
          type: string
        treatmentMade:
          type: string
...